+++ client/src/app/App.js
@@ -27,38 +27,32 @@ import Profile from "../user/profile/Profile";
 import Logs from "../user/logs/Logs";
 import Certificates from "../user/certificates/Certificates";
 import NotFound from "../common/NotFound";
-import LoadingIndicator from "../common/LoadingIndicator";
-import { ASSINA_RSSP_BASE_URL, ACCESS_TOKEN, API_BASE_URL } from "../constants";
 import PersonalizedRoute from "../common/PersonalizedRoute";
 import "./App.css";
-import axios from "axios";
 import AuthProvider from "../common/AuthProviderFunction";
 
 class App extends Component {
-    render() {
-        return (
-            <AuthProvider>
-                <AppHeader />
-                <Routes>
-                    <Route exact path="/" element={<Home />}></Route>
-                    <Route path="/login" element={<Login />}></Route>
-                    <Route path="/redirect" element={<Redirect />}></Route>
-                    <Route path="/sign/redirect" element={<RedirectSign />}></Route>
-                    <Route element={<PersonalizedRoute />}>
-                        <Route path="/profile" element={<Profile />} />
-                        <Route path="/sign" element={<Sign />} />
-                        <Route path="/download" element={<Download />} />
-                        <Route path="/logs" element={<Logs />} />
-                        <Route
-                            path="/certificates"
-                            element={<Certificates />}
-                        />
-                    </Route>
-                    <Route component={NotFound}></Route>
-                </Routes>
-            </AuthProvider>
-        );
-    }
+  render() {
+    return (
+      <AuthProvider>
+        <AppHeader />
+        <Routes>
+          <Route exact path="/" element={<Home />}></Route>
+          <Route path="/login" element={<Login />}></Route>
+          <Route path="/redirect" element={<Redirect />}></Route>
+          <Route path="/sign/redirect" element={<RedirectSign />}></Route>
+          <Route element={<PersonalizedRoute />}>
+            <Route path="/profile" element={<Profile />} />
+            <Route path="/sign" element={<Sign />} />
+            <Route path="/download" element={<Download />} />
+            <Route path="/logs" element={<Logs />} />
+            <Route path="/certificates" element={<Certificates />} />
+          </Route>
+          <Route component={NotFound}></Route>
+        </Routes>
+      </AuthProvider>
+    );
+  }
 }
 
 export default App;