+++ changelog.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [0.5.0]
+_15 Oct 2025_
+
+### Changed
+- Updated OID4VP-related code to support OpenID4VP v1.
+- Updated HSM key templates.
+
+### Fixed
+- Fixed a bug in the client service (React.js) that caused the header to display incorrect options for unauthenticated users.
+
 ## [0.4.0]
 
 _07 Jul 2025_